requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
sib-api-v3-sdk==7.6.0
Werkzeug==3.0.1
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9 